requirements.txt
spotipy
pthon-dotenv